+++ Libraries/elf_loader/include/private/elf_symbol.h
@@ -12,9 +12,7 @@
 extern "C" {
 #endif
 
-// Tactility custom -->
-#define ESP_ELFSYM_EXPORT(_sym)     { #_sym, (const void*)&_sym }
-// <-- Tactility custom
+#define ESP_ELFSYM_EXPORT(_sym)     { #_sym, (void*)&_sym }
 #define ESP_ELFSYM_END              { NULL,  NULL }
 
 /** @brief Function symbol description */
@@ -33,9 +31,23 @@ struct esp_elfsym {
  */
 uintptr_t elf_find_sym(const char *sym_name);
 
-// Tactility custom -->
-void elf_set_custom_symbols(const struct esp_elfsym* symbols);
-// <-- Tactility custom
+
+/**
+ * @brief Resolves a symbol name (e.g. function name) to its address.
+ *
+ * @param sym_name - Symbol name
+ * @return Symbol address if success or 0 if failed.
+ */
+typedef uintptr_t (*symbol_resolver)(const char *sym_name);
+
+/**
+ * @brief Override the internal symbol resolver.
+ * The default resolver is based on static lists that are determined by KConfig.
+ * This override allows for an arbitrary implementation.
+ *
+ * @param resolver the resolver function
+ */
+void elf_set_symbol_resolver(symbol_resolver resolver);
 
 #ifdef __cplusplus
 }