+++ Libraries/elf_loader/CHANGELOG.md
@@ -1,5 +1,16 @@
 # ChangeLog
 
+## v1.1.1 - 2025-06-26
+
+* Added support for ESP32-C61
+
+## v1.1.0 - 2025-05-06
+
+* Added fast build for ELF application
+* Added a script to generate the symbol table for the ELF APP:
+  * Supports generating symbols table based on ELF file
+  * Supports generating symbols table based on static libraries
+
 ## v1.0.0 - 2024-12-09
 
 * Added support for the following RISC-V chips: ESP32-P4 and ESP32-C6